+++ runtime/onert/core/include/ir/Graph.h
@@ -27,17 +27,6 @@
 
 namespace onert
 {
-namespace backend
-{
-namespace custom
-{
-class IKernelBuilder;
-} // namespace custom
-} // namespace backend
-} // namespace onert
-
-namespace onert
-{
 namespace ir
 {
 
@@ -112,22 +101,6 @@ private:
   // TODO Rename to `sweepUnusedOperands`
   // TODO Make this public
   void sweepGarbageOperands();
-
-  // Custom operations support
-public:
-  void
-  bindKernelBuilder(const std::shared_ptr<onert::backend::custom::IKernelBuilder> &kernel_builder)
-  {
-    _kernel_builder = kernel_builder;
-  }
-
-  const std::shared_ptr<backend::custom::IKernelBuilder> &getKernelBuilder() const
-  {
-    return _kernel_builder;
-  }
-
-private:
-  std::shared_ptr<backend::custom::IKernelBuilder> _kernel_builder;
 
   // Accessors
 public:
