+++ runtime/onert/core/src/compiler/ExecutorFactory.h
@@ -37,6 +37,7 @@ struct ExecutorFactoryArgs
   const util::TracingCtx *tracing_ctx;
   const compiler::CompilerOptions *options;
   ir::ModelIndex model_index;
+  std::shared_ptr<backend::custom::IKernelBuilder> custom_kernel_builder;
 };
 
 class ExecutorFactory
